refactor(types): split public types by domain
Keep src/types/index.ts as the stable public barrel while moving the large type surface into focused domain files. This makes the API contracts easier to review without changing existing imports.
