+++ src/types/options.ts
@@ -1,0 +1,2 @@
+export type { ProcessDocumentOptions } from './processDocumentOptions.js';
+export type { ProcessOptions } from './processOptions.js';
