+++ src/types/page.ts
@@ -1,0 +1,2 @@
+export type { PageOverview } from './pageOverview.js';
+export type { PageResult } from './pageResult.js';
